Extract fixture to a common fixture location
